requirements.txt
albumentations==1.0.3
certifi==2022.9.14
click==8.1.3
cycler==0.11.0
fastapi==0.85.0
fonttools==4.37.3
h11==0.14.0
idna==3.4
imageio==2.22.0
importlib-metadata==4.12.0
imutils==0.5.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.5.3
networkx==2.6.3
numpy==1.21.6
opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==9.2.0
pydantic==1.9.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
PyWavelets==1.3.0
PyYAML==6.0
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.3
setuptools==63.4.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.8.0
torchvision== 0.9.0
tqdm==4.62.2
uvicorn==0.18.3
seaborn==0.12.0
opencv-contrib-python==4.6.0.66